refactor(workspace-refactor): split rename_symbol into SRP submodule
Extract the 135-line `WorkspaceRefactor::rename_symbol` body into a new
`workspace_refactor/rename.rs` submodule with one helper per phase:

- `validate_inputs` — input guards (empty / identical names)
- `build_target` — derive sigil, kind, and SymbolKey from old_name
- `collect_indexed_locations` — refs + definition via workspace index
- `fallback_text_search` — bounded naive byte scan when index is empty
- `build_file_edits` — locations -> per-file TextEdits with sigil preservation
- `replacement_text` — compute replacement string for one occurrence

The public method becomes a thin orchestrator (~20 lines) that reads
top-down. Behavior is unchanged; the two fallback-search soft caps that
were inline magic numbers (1000 total, 500 per-doc early-exit) are now
named constants. All 371 existing tests in the crate continue to pass.
